精简命令行进度

+++ hongs-core/src/main/java/io/github/ihongs/combat/CombatHelper.java
@@ -494,7 +494,7 @@ public class CombatHelper
     }
     else
     {
-      CombatHelper.progres ( 0.0f, "..." );
+      CombatHelper.progres ((float) 0 , "..." );
     }
   }
